langchain_pipeline/requirements.txt
langchain
langchain-openai
langchain-anthropic
langchain-community
pydantic
pandas
pypdf
openpyxl
python-dotenv
